… dev_docs/ARCHITECTURE.md
… internal/beater/server_test.go
… internal/model/modelindexer/indexer_test.go
… systemtest/fleet_test.go