@marclop 2022-10-18
## Benchstat diff

### 1 to 4g APM Server

All had similar performance to main, since the active indexers were only scaled up briefly when CPU credits were available which to significant bursts in throughput when that occurred.

### 8g APM Server

This one has less of a noticeable increase, but the CPU was utilized more fully. Bursts had way higher throughput.

```console
$ benchstat -alpha 0.11 sizes/main-active/8g-12s-12n.txt sizes/scaling/8g-24s-24n1mb-50av-gomaxprocs-by-4.txt
name          old time/op                  new time/op                  delta
AgentAll-512                   606ms ± 1%                   593ms ±25%      ~     (p=0.700 n=3+3)

name          old error_responses/sec      new error_responses/sec      delta
AgentAll-512                    0.03 ±58%                    0.00       -100.00%  (p=0.100 n=3+3)

name          old events/sec               new events/sec               delta
AgentAll-512                   29.5k ± 1%                   31.5k ±28%      ~     (p=0.700 n=3+3)

name          old gc_cycles                new gc_cycles                delta
AgentAll-512                     318 ± 1%                     312 ±26%      ~     (p=0.700 n=3+3)

name          old max_goroutines           new max_goroutines           delta
AgentAll-512                     382 ± 2%                     398 ±29%      ~     (p=0.700 n=3+3)

name          old max_heap_alloc           new max_heap_alloc           delta
AgentAll-512                   1.11G ± 2%                   1.15G ± 5%      ~     (p=0.200 n=3+3)

name          old max_heap_objects         new max_heap_objects         delta
AgentAll-512                   10.0M ± 2%                   10.1M ± 6%      ~     (p=1.000 n=3+3)

name          old max_rss                  new max_rss                  delta
AgentAll-512                   1.22G ± 1%                   1.26G ± 5%      ~     (p=0.400 n=3+3)

name          old mean_available_indexers  new mean_available_indexers  delta
AgentAll-512                    16.1 ± 1%                    40.7 ± 8%  +153.35%  (p=0.100 n=3+3)

name          old alloc/op                 new alloc/op                 delta
AgentAll-512                   570MB ± 1%                   581MB ± 1%    +2.03%  (p=0.100 n=3+3)

name          old allocs/op                new allocs/op                delta
AgentAll-512                   7.99M ± 1%                   8.15M ± 1%    +2.09%  (p=0.100 n=3+3)
```

### 15g APM Server

From this size onwards is where the autoscaling really shines. CPU utilization was increased and so was throughput.

```console
$ benchstat -alpha 0.11 sizes/main-active/15g-12s-12n.txt sizes/scaling/15g-24s-24n1mb-50av-gomaxprocs-by-4.txt
name          old time/op                  new time/op                  delta
AgentAll-960                   561ms ±10%                   336ms ± 0%   -40.08%  (p=0.100 n=3+3)

name          old error_responses/sec      new error_responses/sec      delta
AgentAll-960                    0.06 ±85%                   0.00 ±200%   -95.79%  (p=0.100 n=3+3)

name          old events/sec               new events/sec               delta
AgentAll-960                   31.8k ±10%                   53.6k ± 0%   +68.34%  (p=0.100 n=3+3)

name          old gc_cycles                new gc_cycles                delta
AgentAll-960                     317 ± 3%                     448 ± 3%   +41.28%  (p=0.100 n=3+3)

name          old max_goroutines           new max_goroutines           delta
AgentAll-960                     386 ± 3%                     562 ± 2%   +45.64%  (p=0.100 n=3+3)

name          old max_heap_alloc           new max_heap_alloc           delta
AgentAll-960                   1.11G ± 0%                   1.19G ± 1%    +7.72%  (p=0.100 n=3+3)

name          old max_heap_objects         new max_heap_objects         delta
AgentAll-960                   9.89M ± 1%                  10.67M ± 0%    +7.91%  (p=0.100 n=3+3)

name          old max_rss                  new max_rss                  delta
AgentAll-960                   1.23G ± 2%                   1.33G ± 1%    +7.84%  (p=0.100 n=3+3)

name          old mean_available_indexers  new mean_available_indexers  delta
AgentAll-960                    15.5 ± 7%                    33.4 ± 0%  +115.79%  (p=0.100 n=3+3)

name          old alloc/op                 new alloc/op                 delta
AgentAll-960                   564MB ± 1%                   571MB ± 0%    +1.20%  (p=0.100 n=3+3)

name          old allocs/op                new allocs/op                delta
AgentAll-960                   7.92M ± 1%                   8.07M ± 0%    +1.82%  (p=0.100 n=3+3)
```

### 30g APM Server

```console
$ benchstat -alpha 0.11 sizes/main-active/30g-12s-12n.txt sizes/scaling/30g-24s-24n1mb-50av-gomaxprocs-by-4-pr.txt
name           old time/op                  new time/op                  delta
AgentAll-1920                   585ms ± 5%                   183ms ± 0%   -68.71%  (p=0.100 n=3+3)

name           old error_responses/sec      new error_responses/sec      delta
AgentAll-1920                    0.04 ±18%                    0.01 ±50%   -76.61%  (p=0.100 n=3+3)

name           old events/sec               new events/sec               delta
AgentAll-1920                   30.7k ± 5%                   98.1k ± 0%  +219.84%  (p=0.100 n=3+3)

name           old gc_cycles                new gc_cycles                delta
AgentAll-1920                     335 ± 5%                     772 ± 1%  +130.55%  (p=0.100 n=3+3)

name           old max_goroutines           new max_goroutines           delta
AgentAll-1920                     390 ± 2%                    1026 ± 1%  +163.22%  (p=0.100 n=3+3)

name           old max_heap_alloc           new max_heap_alloc           delta
AgentAll-1920                   1.09G ± 2%                   1.41G ± 1%   +29.59%  (p=0.100 n=3+3)

name           old max_heap_objects         new max_heap_objects         delta
AgentAll-1920                   9.18M ± 3%                  13.15M ± 1%   +43.31%  (p=0.100 n=3+3)

name           old max_rss                  new max_rss                  delta
AgentAll-1920                   1.23G ± 2%                   1.58G ± 0%   +28.20%  (p=0.100 n=3+3)

name           old mean_available_indexers  new mean_available_indexers  delta
AgentAll-1920                    15.8 ± 3%                    12.2 ± 7%   -22.64%  (p=0.100 n=3+3)

name           old alloc/op                 new alloc/op                 delta
AgentAll-1920                   573MB ± 0%                   560MB ± 0%    -2.19%  (p=0.100 n=3+3)

name           old allocs/op                new allocs/op                delta
AgentAll-1920                   8.00M ± 0%                   7.96M ± 0%    -0.59%  (p=0.100 n=3+3)
```