@marclop 2022-10-18
## Metric Screenshots

The last 3 distributions are all for the same 30g APM Server instance. I ran the benchmarks 3 times with that size to gather some more data.

## APM Server

<img width="1545" alt="Screen Shot 2022-10-18 at 15 57 11" src="https://user-images.githubusercontent.com/7286993/196381551-2767a10f-ead2-4cbe-91b6-1a0d385585b0.png">

## Elasticsearch

464GB of RAM! Cluster over 3 zones. With a total of 24 hot nodes.

<img width="1687" alt="Screen Shot 2022-10-18 at 16 06 53" src="https://user-images.githubusercontent.com/7286993/196381734-d43e115f-cc22-4d29-96d2-dbbba85430aa.png">

<img width="1687" alt="Screen Shot 2022-10-18 at 16 07 23" src="https://user-images.githubusercontent.com/7286993/196381763-28bd6eb8-743a-41ba-8bd5-e67e2214ce32.png">

<img width="1687" alt="Screen Shot 2022-10-18 at 16 08 24" src="https://user-images.githubusercontent.com/7286993/196381814-e0da6d54-925d-4e4e-817e-f962bbe0d302.png">